+++ all-is-cubes-mesh/src/dynamic/chunked_mesh/tests.rs
@@ -47,11 +47,11 @@ fn update_adjacent_chunk_positive() {
         (ChunkPos::new(0, 0, 0), ChunkTodo::CLEAN),
         (ChunkPos::new(1, 0, 0), ChunkTodo::CLEAN),
     ]);
-    listener.receive(SpaceChange::CubeBlock {
+    listener.receive(&[SpaceChange::CubeBlock {
         cube: Cube::new(CHUNK_SIZE - 1, CHUNK_SIZE / 2, CHUNK_SIZE / 2),
         old_block_index: 123,
         new_block_index: 456,
-    });
+    }]);
     assert_eq!(
         read_todo_chunks(&todo),
         vec![
@@ -83,11 +83,11 @@ fn update_adjacent_chunk_negative() {
         (ChunkPos::new(0, 0, 0), ChunkTodo::CLEAN),
         (ChunkPos::new(1, 0, 0), ChunkTodo::CLEAN),
     ]);
-    listener.receive(SpaceChange::CubeBlock {
+    listener.receive(&[SpaceChange::CubeBlock {
         cube: Cube::new(0, CHUNK_SIZE / 2, CHUNK_SIZE / 2),
         old_block_index: 123,
         new_block_index: 456,
-    });
+    }]);
     assert_eq!(
         read_todo_chunks(&todo),
         vec![
@@ -117,22 +117,22 @@ fn todo_ignores_absent_chunks() {
 
     let cube = Cube::from(GridPoint::new(1, 1, 1) * (CHUNK_SIZE / 2));
     // Nothing happens...
-    listener.receive(SpaceChange::CubeBlock {
+    listener.receive(&[SpaceChange::CubeBlock {
         cube,
         old_block_index: 0,
         new_block_index: 0,
-    });
+    }]);
     assert_eq!(read_todo_chunks(&todo), vec![]);
     // until the chunk exists in the table already.
     todo.lock()
         .unwrap()
         .chunks
         .insert(ChunkPos::new(0, 0, 0), ChunkTodo::CLEAN);
-    listener.receive(SpaceChange::CubeBlock {
+    listener.receive(&[SpaceChange::CubeBlock {
         cube,
         old_block_index: 0,
         new_block_index: 0,
-    });
+    }]);
     assert_eq!(
         read_todo_chunks(&todo),
         vec![(